WIP : [IMP] runbot: add waiting jobs
It will now be possible to wait for event before a build is finished
or execute next step.
For that, we can add a wait_children step or add the flag
'waiting_job' on a python or create step.

Waiting step are not counted in used runbot slots, so it is possible
to go over the max, the rule is to only use light step
after a waiting step (cla check, make result, ...) no docker run.