[FIX] runbot: github status should be based on global_state of top parent
If to_parent is a special_case, his config will have update_github_state False
avoiding to contact github in this case.

If a child wants to contact github for failfast, parent will be in global_state error too.
If a child want to contact github for succsess, we actually want to wait for last child.
Only last child will be able to notiffy success since global_state will be running or done at this step.
Orphan builds wont have any impact on result in with this scenario.